Swing Interface and Fixer API implemented

+++ src/main/java/software/eii/ulpgc/psl/moneycalculator/ExchangeRate.java
@@ -2,5 +2,5 @@ package software.eii.ulpgc.psl.moneycalculator;
 
 import java.time.LocalDate;
 
-public record ExchangeRate(Currency from, Currency to, long rate, LocalDate date){
+public record ExchangeRate(Currency from, Currency to, double rate, LocalDate date){
 }
