+++ src/main/java/software/eii/ulpgc/psl/moneycalculator/Money.java
@@ -1,6 +1,6 @@
 package software.eii.ulpgc.psl.moneycalculator;
 
-public record Money(Currency currency, long quantity) {
+public record Money(Currency currency, double quantity) {
     @Override
     public String toString() {
         return quantity + " " + currency.code();
